Add update/fetch func, expiration check for Note
- Adds has_expired property for Note objects that checks for date
expiration
- Adds should_not_exist property for Note objects that checks for date
expiration and visits count validation
- Adds fetch function for note ids which verifies that the Note
should exist. This will ensure that notes that have expired due to date
validation, should not be fetched. If such validation fails, note is
deleted and error is raised
- Adds update function for note objects that updates the visit counter
and checks if there can be another visit to that Note. If for example a
note should be visited only once, update_note will increase the visits
counter and then it will check that the counter is the maximum that is
allowed and will delete the object
- Adds tests for models module