Trigger update_note after fetching payload
- Calls `update_note` when fetching note, after storing the note
payload. This will ensure that if a note should be deleted after
visit, it will be deleted even if it cannot be decrypted. For
paranoid explanation, this solves scenarios where someone could
"bruteforce" the decryption of a known payload stored in the
database. This only solves scenarios for stored payloads that are
bruteforced through web requests. It does not solve scenarios where
the payload is embeded in the url (aka no-store option) or cases
where the adversary holds control of the database (and thus can try
bruteforcing or any other attack on a stored payload)