Fix bugs with key derivation functionality
1. Fixes bug with key derivation function where salt was reused for
multiple key derivations (due to salt being defined in the module and
not inside a function)
2. Fixes bug where there was the assumption that through the password you
could generate the same key. This one slipped because due to [1]. Since the
salt was generated in the module, this allowed tests to craft the same
key with a password. Testing real life (with server reboots) and fixing
[1] uncovered this well hidden bug. So now there's no
decrypt_with_password function since we don't provide a return of the
salt. In future we could provide the salt as well if we wanted to
provide support for description server agnostic. Aka if you know the
password, the salt and the ciphertext you can even decrypt this on your
own computer with pynacl. Currently out of scope.