Add new config option for enc algorithm parameters
- Adds new config option SUPPORTED_ALGORITHM_PARAMS to support choosing
of encryption algorithm and hardness when encrypting a note. Options
must be a specificly formatted dictionary that will be parsed as json
and will be validated to see if the given options are valid before
starting the server (i.e if admin provided wrongly formatted options or
adds an algorithm/hardness that is not supported. Argon2i is not
supported by default on all machines)